More changes for generic database, WIP artifact record

+++ src/org/lasalledebain/libris/Schema.java
@@ -22,7 +22,7 @@ public abstract class Schema implements LibrisXMLConstants, XMLElement {
 	public static String currentVersion = "1.0";
 	protected TreeMap<String, EnumFieldChoices> enumSets;
 	protected ArrayList<FieldTemplate> fieldList;
-	HashMap <String, Short> fieldNumById;
+	HashMap<String, Integer> fieldNumById;
 	ArrayList<String> fieldIds;
 	ArrayList<String> fieldTitles;
 	String fieldIdArray[];
@@ -34,7 +34,7 @@ public abstract class Schema implements LibrisXMLConstants, XMLElement {
 	public Schema() {
 		enumSets = new TreeMap<String, EnumFieldChoices>();
 		fieldList = new ArrayList<FieldTemplate>();
-		fieldNumById = new HashMap<String, Short>();
+		fieldNumById = new HashMap<String, Integer>();
 		fieldIds = new ArrayList<String>();
 		fieldTitles = new ArrayList<String>();
 	}
@@ -76,7 +76,7 @@ public abstract class Schema implements LibrisXMLConstants, XMLElement {
 		String fieldId = field.getFieldId();
 		fieldIds.add(fieldId);
 		fieldTitles.add(field.getFieldTitle());
-		fieldNumById.put(fieldId, (short) (fieldIds.size()-1));
+		fieldNumById.put(fieldId, (fieldIds.size()-1));
 		final int index = fieldIds.size()-1;
 		fieldList.add(index, field);
 		return index;
@@ -123,12 +123,12 @@ public abstract class Schema implements LibrisXMLConstants, XMLElement {
 	 * @param fieldId string name of the field
 	 * @return index of field in the list of fields, or -1 if the field is not found.
 	 */
-	public short getFieldNum(String fieldId) {
-		Short fieldIndex = fieldNumById.get(fieldId);
+	public int getFieldNum(String fieldId) {
+		Integer fieldIndex = fieldNumById.get(fieldId);
 		if (null == fieldIndex) {
 			return LibrisConstants.NULL_FIELD_NUM;
 		} else {
-			return fieldIndex.shortValue();
+			return fieldIndex.intValue();
 		}
 	}
 
@@ -145,7 +145,7 @@ public abstract class Schema implements LibrisXMLConstants, XMLElement {
 	}
 
 	public FieldType getFieldType(String i) {
-		Short fieldNum = fieldNumById.get(i);
+		Integer fieldNum = fieldNumById.get(i);
 		FieldTemplate f = fieldList.get(fieldNum);
 		FieldType t = f.getFtype();
 		return t;
